requirements.txt
alembic==1.10.4
aniso8601==9.0.1
appdata==2.1.2
astroid==2.15.5
attrs==23.1.0
blinker==1.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
colored==1.4.3
cryptography==40.0.2
dill==0.3.6
exceptiongroup==1.1.1
Flask==2.3.2
flask-marshmallow==0.15.0
Flask-Migrate==2.7.0
flask-restx==1.1.0
Flask-Script==2.0.5
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
packaging==23.1
pep8==1.7.1
platformdirs==3.5.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
PyJWT==1.7.1
pylint==2.17.4
pylint-flask==0.6
pylint-plugin-utils==0.8.1
pyrsistent==0.19.3
pytest==7.3.1
pytest-flask==1.2.0
pytest-helper==0.2.2
python-dotenv==1.0.0
pytz==2023.3
set-package-attribute==0.2.5
SQLAlchemy==2.0.13
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
Werkzeug==2.3.4
wrapt==1.15.0